requirements.txt
flask
google-generativeai
gunicorn
pysqlite3
markdown